Improved way of getting a buffers client node
